Added controls to skip high computing steps

+++ bin/coPatReporter.rb
@@ -284,8 +284,8 @@ write_arrays4scatterplot(onto_ic.values, freq_ic.values, hpo_ic_file, 'OntoIC', 
 write_arrays4scatterplot(onto_ic_profile.values, freq_ic_profile.values, hpo_profile_ic_file, 'OntoIC', 'FreqIC') #HP profiles
 write_arrays4scatterplot(profile_sizes, parental_hpos_per_profile, parents_per_term_file, 'ProfileSize', 'ParentTerms')
 
-system("#{File.join(EXTERNAL_CODE, 'plot_scatterplot_simple.R')} -i #{hpo_ic_file} -o #{File.join(temp_folder, 'hpo_ics.pdf')} -x 'OntoIC' -y 'FreqIC' --x_tag 'HP Ontology IC' --y_tag 'HP Frequency based IC' --x_lim '0,4.5' --y_lim '0,4.5'")
-system("#{File.join(EXTERNAL_CODE, 'plot_scatterplot_simple.R')} -i #{hpo_profile_ic_file} -o #{File.join(temp_folder, 'hpo_profile_ics.pdf')} -x 'OntoIC' -y 'FreqIC' --x_tag 'HP Ontology Profile IC' --y_tag 'HP Frequency based Profile IC' --x_lim '0,4.5' --y_lim '0,4.5'")
+system("#{File.join(EXTERNAL_CODE, 'plot_scatterplot_simple.R')} -i #{hpo_ic_file} -o #{File.join(temp_folder, 'hpo_ics.pdf')} -x 'OntoIC' -y 'FreqIC' --x_tag 'HP Ontology IC' --y_tag 'HP Frequency based IC' --x_lim '0,4.5' --y_lim '0,4.5'") if !File.exists?(File.join(temp_folder, 'hpo_ics.pdf'))
+system("#{File.join(EXTERNAL_CODE, 'plot_scatterplot_simple.R')} -i #{hpo_profile_ic_file} -o #{File.join(temp_folder, 'hpo_profile_ics.pdf')} -x 'OntoIC' -y 'FreqIC' --x_tag 'HP Ontology Profile IC' --y_tag 'HP Frequency based Profile IC' --x_lim '0,4.5' --y_lim '0,4.5'") if !File.exists?(File.join(temp_folder, 'hpo_profile_ics.pdf'))
 system("#{File.join(EXTERNAL_CODE, 'plot_scatterplot_simple.R')} -i #{parents_per_term_file} -o #{File.join(temp_folder, 'parents_per_term.pdf')} -x 'ProfileSize' -y 'ParentTerms' --x_tag 'Patient HPO profile size' --y_tag 'Parent HPO terms within the profile'")
 
 ###Cohort frequency calculation
@@ -314,24 +314,28 @@ end
 #----------------------------------
 Parallel.each(options[:clustering_methods], in_processes: options[:threads] ) do |method_name|
   matrix_filename = File.join(temp_folder, "similarity_matrix_#{method_name}.npy")
+  axis_file = matrix_filename.gsub('.npy','.lst')
   profiles_similarity_filename = File.join(temp_folder, ['profiles_similarity', method_name].join('_').concat('.txt'))
   clusters_distribution_filename = File.join(temp_folder, ['clusters_distribution', method_name].join('_').concat('.txt'))
-  profiles_similarity = hpo.compare_profiles(sim_type: method_name.to_sym)
-  profile_pairs = write_profile_pairs(profiles_similarity, profiles_similarity_filename)
-  similarity_matrix, axis_names = format_profiles_similarity_data_numo(profiles_similarity)
-  axis_file = matrix_filename.gsub('.npy','.lst')
-  File.open(axis_file, 'w'){|f| f.print axis_names.join("\n") }
-  Npy.save(matrix_filename, similarity_matrix)
+  if !File.exists?(matrix_filename)
+    profiles_similarity = hpo.compare_profiles(sim_type: method_name.to_sym)
+    write_profile_pairs(profiles_similarity, profiles_similarity_filename)
+    similarity_matrix, axis_names = format_profiles_similarity_data_numo(profiles_similarity)
+    File.open(axis_file, 'w'){|f| f.print axis_names.join("\n") }
+    Npy.save(matrix_filename, similarity_matrix)
+  end
   ext_var = ''
   if method_name == 'resnik'
     ext_var = '-m max'
   elsif method_name == 'lin'
     ext_var = '-m comp1'
   end
-  system("#{File.join(EXTERNAL_CODE, 'plot_heatmap.R')} -y #{axis_file} -d #{matrix_filename} -o #{File.join(temp_folder, method_name)} -H #{ext_var}")
+  out_file = File.join(temp_folder, method_name)
+  system("#{File.join(EXTERNAL_CODE, 'plot_heatmap.R')} -y #{axis_file} -d #{matrix_filename} -o #{out_file} -H #{ext_var}") if !File.exists?(out_file +  '_heatmap.png')
   clusters_codes, clusters_info = parse_clusters_file(File.join(temp_folder, "#{method_name}_clusters.txt"), patient_uniq_profiles)
   get_cluster_metadata(clusters_info, clusters_distribution_filename)
-  system("#{File.join(EXTERNAL_CODE, 'xyplot_graph.R')} -d #{clusters_distribution_filename} -o #{File.join(temp_folder, ['clusters_distribution', method_name].join('_'))} -x PatientsNumber -y HPOAverage")
+  out_file = File.join(temp_folder, ['clusters_distribution', method_name].join('_'))
+  system("#{File.join(EXTERNAL_CODE, 'xyplot_graph.R')} -d #{clusters_distribution_filename} -o #{out_file} -x PatientsNumber -y HPOAverage") if !File.exists?(out_file)
   clusters = translate_codes(clusters_codes, hpo)
   
   container = {
